+++ src/api/fetch/auth/api/useApiFindPw.ts
@@ -3,12 +3,10 @@ import { ApiBaseResponseType } from "@/api/_base/types/ApiBaseResponseType";
 import { ApiFindPwType } from "../types/ApiFindPwType";
 import { AxiosError } from "axios";
 
-const useApiFindPw = () => {
+export const useApiFindPw = () => {
   return useAppMutation<
     ApiFindPwType,
     ApiBaseResponseType<null>,
     AxiosError<ApiBaseResponseType<null>>
   >("public", "/auth/reset/request", "post");
 };
-
-export default useApiFindPw;
