+++ src/api/fetch/auth/api/useApiLogout.ts
@@ -1,12 +1,10 @@
 import useAppMutation from "@/api/_base/query/useAppMutation";
 import { ApiBaseResponseType } from "@/api/_base/types/ApiBaseResponseType";
 
-const useApiLogout = () => {
+export const useApiLogout = () => {
   return useAppMutation<void, ApiBaseResponseType<null>, ApiBaseResponseType<null>>(
     "auth",
     "/auth/logout",
     "post"
   );
 };
-
-export default useApiLogout;
