+++ src/api/fetch/auth/api/useApiSignUp.ts
@@ -4,12 +4,10 @@ import { SignUpResponseType } from "../types/SignUpResponseType";
 import { ApiBaseResponseType } from "@/api/_base/types/ApiBaseResponseType";
 import { AxiosError } from "axios";
 
-const useApiSignUp = () => {
+export const useApiSignUp = () => {
   return useAppMutation<ApiSignUpType, SignUpResponseType, AxiosError<ApiBaseResponseType<null>>>(
     "auth",
     "auth/signup",
     "post"
   );
 };
-
-export default useApiSignUp;
