+++ src/api/fetch/inquiry/index.ts
@@ -5,3 +5,4 @@ export * from "./types/GetByIdInquiryResponseType";
 export { usePostInquiry } from "./api/usePostInquiry";
 export { useGetUserInquiries } from "./api/useGetUserInquiries";
 export { useGetUserInquiryById } from "./api/useGetUserInquiryById";
+export { usePostUserInquiry } from "./api/usePostUserInquiry";
