+++ src/api/fetch/notification/api/useGetNotificationSetting.ts
@@ -1,7 +1,7 @@
 import useAppQuery from "@/api/_base/query/useAppQuery";
 import { GetNotificationSetting } from "../types/notificationSettingType";
 
-const useGetNotificationSetting = (options?: { enabled?: boolean }) => {
+export const useGetNotificationSetting = (options?: { enabled?: boolean }) => {
   return useAppQuery<GetNotificationSetting, unknown>(
     "auth",
     ["/notifications/settings"],
@@ -9,5 +9,3 @@ const useGetNotificationSetting = (options?: { enabled?: boolean }) => {
     options
   );
 };
-
-export default useGetNotificationSetting;
