+++ src/api/fetch/notification/index.ts
@@ -7,4 +7,5 @@ export { default as useNotificationSSE } from "./api/useNotificationSSE";
 export { default as useNotificationList } from "./api/useNotificationList";
 export { default as useNotificationRead } from "./api/useNotificationRead";
 export { default as useNotificationDeleteAll } from "./api/useNotificationDeleteAll";
-export { default as useGetNotificationSetting } from "./api/useGetNotificationSetting";
+export { useGetNotificationSetting } from "./api/useGetNotificationSetting";
+export { usePutNotificationSetting } from "./api/usePutNotificationSetting";
