+++ src/api/fetch/report/index.ts
@@ -3,9 +3,9 @@ export * from "./types/ReportRequest";
 export * from "./types/GetReportsResponseType";
 export * from "./types/GetReportsByIdResponseType";
 
-export { default as useGetBlockUser } from "./api/useGetBlockUser";
-export { default as useBlock } from "./api/useBlock";
-export { default as useReport } from "./api/useReport";
-export { default as useDeleteBlockUser } from "./api/useDeleteBlockUser";
-export { default as useGetUserReports } from "./api/useGetUserReports";
-export { default as useGetReportById } from "./api/useGetReportById";
+export { useGetBlockUser } from "./api/useGetBlockUser";
+export { useBlock } from "./api/useBlock";
+export { useReport } from "./api/useReport";
+export { useDeleteBlockUser } from "./api/useDeleteBlockUser";
+export { useGetUserReports } from "./api/useGetUserReports";
+export { useGetReportById } from "./api/useGetReportById";
