+++ src/components/domain/Terms/Terms.tsx
@@ -4,8 +4,7 @@ import { DetailHeader } from "@/components/layout";
 import { FooterButton } from "@/components/domain";
 import { CheckBox } from "@/components/common";
 import { TERM_CONTENTS } from "./_constants/TERM_CONTENTS";
-import { useGetNotificationSetting } from "@/api/fetch/notification";
-import usePutNotificationSetting from "@/api/fetch/notification/api/usePutNotificationSetting";
+import { useGetNotificationSetting, usePutNotificationSetting } from "@/api/fetch/notification";
 import { DEFAULT_NOTIFICATION_SETTING } from "@/app/(route)/mypage/notifications/_constants/DEFAULT_NOTIFICATION_SETTING";
 import { useToast } from "@/context/ToastContext";
